backend/requirements.txt
fastapi
uvicorn[standard]
httpx
groq
python-dotenv
sqlalchemy
python-multipart
pydantic[email]
twilio
